requirements.txt
fastapi
sqlalchemy
psycopg2-binary
uvicorn
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pytz
